requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.3.0
black==18.9b0
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==3.2
discord.py==1.3.4
dnspython==1.16.0
email-validator==1.1.0
idna==2.9
ipcalc==1.99.0
multidict==4.7.5
pycparser==2.20
pymongo==3.10.1
six==1.14.0
toml==0.10.0
websockets==8.1
yarl==1.4.2
